LayoutTests/fast/forms/alternative-presentation-button/replaced-elements.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id="test-container">
    <div class="test" data-test-name="single input element">
        <input type="text">
    </div>
    <div class="test" data-test-name="multiple elements">
        <label for="first-name">First name</label>
        <input type="text" id="first-name">
        <label for="last-name">Last name</label>
        <input type="text" id="last-name">
    </div>
</div>
<script>
function areArraysEqualIgnoringOrder(a, b)
{
    if (a.length != b.length)
        return false;
    for (let aElement of a) {
        if (!b.includes(aElement))
            return false;
    }
    return true;
}

function runTests()
{
    if (!window.internals) {
        testFailed("Must be run in DumpRenderTree or WebKitTestRunner.");
        return;
    }

    // Non-existent identifier
    shouldBeZero("internals.elementsReplacedByAlternativePresentationButton(1).length");

    var id = 0;
    var testContainer = document.getElementById("test-container");
    var tests = testContainer.getElementsByClassName("test");
    for (let test of tests) {
        var id = ++id;
        var expectedReplacedElements = Array.prototype.slice.call(test.children);
        internals.substituteWithAlternativePresentationButton(test.children, id);
        var actualReplacedElements = internals.elementsReplacedByAlternativePresentationButton(id);
        if (areArraysEqualIgnoringOrder(actualReplacedElements, expectedReplacedElements))
            testPassed(test.dataset.testName);
        else
            testFailed(test.dataset.testName);
    }
}

description("Tests that the elements returned by internals.elementsReplacedByAlternativePresentationButton() are the same as the elements passed to internals.substituteWithAlternativePresentationButton() ignoring order.");
runTests();
</script>
</body>
</html>
